Rename: main to app
