+++ src/index.css
@@ -14,7 +14,6 @@ body {
   font-family: "Pretendard";
   height: 100vh;
   margin: 0;
-  /* width: 100%; */
 }
 
 #root {
